Add missing puzzle6 words to all-words6
See issue #21, there are a couple of puzzles that are impossible since
they are in puzzle-words6, but not in all-words6. These words are
usually quite obscure ('hiatus', 'abdita'), but in some cases are pretty
regular words ('deluxe').

This commit only adds these missing words to all-words6, it would be
better to 'clean' the wordlist such that some of the obscure words
aren't in the puzzle-words6.

I will add a later commit to do the same for puzzle-words, of length 5.